refactor: add data schema for all backend pydantic models
